split scaling.funcs into scaling.funcs_redshift and scaling.funcs_mass
